+++ src/Products/PloneMeeting/tests/PloneMeetingTestCase.py
@@ -432,7 +432,8 @@ class PloneMeetingTestCase(unittest.TestCase, PloneMeetingTestingHelpers):
                  to_sign=False,
                  signed=False,
                  publishable=False,
-                 annexFile=None):
+                 annexFile=None,
+                 scan_id=None):
         '''Adds an annex to p_item.
            If no p_annexType is provided, self.annexFileType is used.
            If no p_annexTitle is specified, the predefined title of the annex type is used.'''
@@ -471,7 +472,8 @@ class PloneMeetingTestCase(unittest.TestCase, PloneMeetingTestingHelpers):
             confidential=confidential,
             to_sign=to_sign,
             signed=signed,
-            publishable=publishable)
+            publishable=publishable,
+            scan_id=scan_id)
         # need to commit the transaction so the stored blob is correct
         # if not done, accessing the blob will raise 'BlobError: Uncommitted changes'
         transaction.commit()